Move handling of all the pretty-table stuff into its own file.
This declutters html2text.go
